pyproject.toml
[project]
name = "Blender-Launcher-V2"
version = "2.4.0"
description = ""
authors = []
dependencies = [
    "pyinstaller>=6.5.0",
    "beautifulsoup4>=4.11.1",
    "urllib3[socks]>=1.26.11",
    "pynput==1.7.7",
    "lxml",
    "python-xlib",
    "pywin32>=303; sys_platform == \"win32\"",
    "distro>=1.8.0",
    "semver>=3.0.2",
    "zstandard>=0.23.0",
    "send2trash>=1.8.3",
    "webdav4==0.10.0",
    "dateparser>=1.2.0",
    "pyside6-essentials==6.8.1",
    "markdown>=3.7",
]
requires-python = "<3.14,>=3.10"
readme = "README.md"
license = { text = "GPL3" }

[project.optional-dependencies]
docs = ["mkdocs", "mkdocs-material>=9.5.10"]
ruff = ["ruff>=0.3.5"]
testing = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
]


[build-system]
build-backend = "pdm.backend"
requires = ["pdm-backend"]

[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.python.org/simple"
verify_ssl = true
name = "pypi"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
extend-ignore = [
  "F401",   # unused-import -- Certain imports need to be explicitly made, even though they are unused
  "E501",   # line-too-long -- Unavoidable in some cases, but it should still be followed everywhere else
  "SIM114", # if-with-same-arms -- This can make the code harder to read and understand in some places
  "SIM108", # if-else-block-instead-of-if-exp -- Same as above, but this is more common
  "RET505", # superfluous-else-return -- makes related stuff harder to read
]
extend-select = [
  "A",     # flake8-builtins
  "ASYNC", # flake8-async
  "UP",    # pyupgrade
  "I",     # isort
  "B",     # flake8-bugbear
  "ICN",   # flake8-import-conventions
  "F",     # pyflakes
  "RET",   # flake8-return
  "SIM",   # flake8-simplify
  "NPY",   # NumPy-specific rules
  "PERF",  # perflint
  "RUF",   # Ruff-specific rules
  "PIE",   # flake8-pie
  "DTZ",   # flake8-datetimez
  "C4",    # flake8-comprehensions
  "E",     # Error
  "EXE",   # flake8-executable
  "TCH",   # flake8-type-checking
  "W",     # pycodestyle
  "FA",    # flake8-future-annotations
  "ISC",   # flake8-implicit-str-concat
  "Q",     # flake8-quotes
  "SLF",   # flake8-self

  #Specific pylint rules
  # "PL",      # Pylint
  "PLR1711", # useless-return
  "PLR1714", # repeated-equality-comparison

]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"source/resources_rc.py" = ["ALL"]

[tool.isort]
profile = "black"

[tool.setuptools]
py-modules = ["source"]

[tool.pytest.ini_options]
pythonpath = "source"